+++ packages/java-edition/src/mcfunction/tree/patchValidator.ts
